pyproject.toml
[project]
name = "fractal-llm"
version = "0.1.0"
description = "Fractal analysis of LLM fine-tuning trainability boundaries"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.0",
    "modal>=1.3.0.post1",
    "numpy>=2.0.0",
    "rich>=14.0.0",
    "simple-parsing>=0.1.7",
    "tqdm>=4.67.0",
    "pandas>=2.2.0",
    "seaborn>=0.13.0",
    "scipy>=1.14.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "torchaudio==2.8.0",
    "tiktoken>=0.11.0",
    "tokenizers>=0.22.0",
    "transformers>=4.47.0",
    "datasets>=3.2.0",
    "accelerate>=1.2.0",
    "wandb>=0.23.1",
    "huggingface-hub>=0.27.0",
    "lm-eval==0.4.2",
    "python-dotenv>=1.0.1",
]

[tool.uv]
# Use PyTorch cu128 index for torch 2.8.0 wheels
# Use unsafe-best-match to prefer PyPI for non-torch packages
extra-index-url = ["https://download.pytorch.org/whl/cu128"]
index-strategy = "unsafe-best-match"
